Bug 1706772 use size_t for RevebAccumulationBuffer read index r=padenot
The accumulate() return value and updateReadIndex() were unused.

Depends on D113624

Differential Revision: https://phabricator.services.mozilla.com/D113625

+++ dom/media/webaudio/blink/ReverbConvolverStage.h
@@ -71,7 +71,7 @@ class ReverbConvolverStage {
   mozilla::UniquePtr<FFTConvolver> m_fftConvolver;
 
   ReverbAccumulationBuffer* m_accumulationBuffer;
-  int m_accumulationReadIndex;
+  size_t m_accumulationReadIndex;
   int m_inputReadIndex;
 
   size_t m_postDelayLength;
